change to call move function instead of reading table context data
