+++ crates/e2e-move-tests/src/tests/string_viewer.data/viewer2/Move.toml
@@ -8,3 +8,4 @@ StringViewer = { local = "../viewer" }
 
 [addresses]
 std =  "0x1"
+publisher = "0x9999"
