+++ crates/types/src/metadata.rs
@@ -11,6 +11,7 @@ pub const EVENT_STRUCT_ATTRIBUTE: &str = "event";
 pub const CODE_MODULE_NAME: &str = "code";
 pub const INIT_MODULE_FUNCTION_NAME: &str = "init_module";
 pub const INIT_GENESIS_FUNCTION_NAME: &str = "init_genesis";
+pub const VERIFY_DEPENDENCIES_UPGRADE_POLICY_FUNCTION_NAME: &str = "verify_dependencies_upgrade_policy";
 pub const METADATA_V0_MIN_FILE_FORMAT_VERSION: u32 = 6;
 
 /// The keys used to identify the metadata in the metadata section of the module bytecode.
